added Multi-Stat files v0.1.1 + Powershell module
Added Multi-Stat server status checker files:
/css/
/images/
/Powershell/minestat.psm1
.htaccess
index.php
settings.php

Minestat powershell module:
Made small (insignificant) changes to the minestat powershell file to "prettify" the outputs on Multi-Stat

* Altered values to make it pretty, not necessary.
"Success" => "Online"
"Unknown" => "Offline"

Removed 'hidden' from most of the output values, changed the ones I did not want showing up to 'hidden' and reordered the values so it outputs in a more appealing order on Multi-Stat

IMPORTANT CHANGE:
formatted_motd is hidden and stripped_motd is in its place.  When outputting the data to PHP, it shows all the color characters as broken/nonexistent symbols, this was very ugly on the web status page.  
THIS IS NOT A NECESSARY CHANGE BUT I'D RECOMMEND IT!!! 


To reference the changes, view the original minestat.psm1 at the link below and search the following:

[Flags()] enum SlpProtocol (changed Unknown to Offline)
class ServerStatus (reordered values, unhid most of them, hid formatted_motd, motd, slp_protocol and favicon)
 
https://github.com/FragLand/minestat/blob/master/PowerShell/MineStat/MineStat.psm1